Autonomous Treasure Hunter
In the Autonomous Treasure Hunter project, it is expected to design a game in which an autonomous moving character collects treasures on a map with various treasures and obstacles. The goal of the game is to design an algorithm that will allow the character to collect all the treasures in the shortest possible time. For this purpose, it is expected to use object-oriented programming and data structures information.